requirements.txt

livekit>=0.19.1
livekit-agents[silero,google]>=1.4.0
livekit-plugins-google
livekit-plugins-silero
livekit-plugins-openai
python-dotenv
livekit-api
flask[async]
flask-cors
uvicorn
livekit-plugins-noise-cancellation~=0.2